% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_key.R
\name{return_key}
\alias{return_key}
\title{Return Setlist FM API Key from environment}
\usage{
return_key()
}
\value{
\code{return_key} returns a key stored on the environment
}
\description{
Return Setlist FM API Key from environment
}
