% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_artists.R
\name{search_artists}
\alias{search_artists}
\alias{search_artist}
\title{Search Artists on SetListFM API.}
\usage{
search_artist(key, search_term, search_by = "artistName")
}
\arguments{
\item{key}{API key.}

\item{search_term}{Artist to be searched}

\item{search_by}{Parameter to be searched - Name, Musicbrainz Identifier, Ticketmaster Identifier}
}
\value{
\code{search_artists} returns a tibble with artists data
}
\description{
Returns artists from SetlistFM API.
}
\examples{

\dontrun{
key <- "your token"
search_term <- "the-beatles"
search_artists(key = key,
search_term = artist)
}


}
\author{
Douglas Zickuhr, \email{douglasrsl@gmail.com}.
}
