% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_venues.R
\docType{data}
\name{search_venues}
\alias{search_venues}
\alias{search_by_options}
\title{Search Venues on SetListFM API.}
\format{An object of class \code{character} of length 6.}
\usage{
search_by_options
}
\arguments{
\item{key}{API key.}

\item{search_term}{Term to be searched}

\item{search_by}{Parameter to be searched}
}
\value{
\code{search_venues} returns a tibble with venues
}
\description{
Returns venues from SetlistFM API.
}
\examples{

\dontrun{
key <- "your token"
artist <- "the-beatles"
get_request(key = key,
artist = artist)
}



}
\author{
Douglas Zickuhr, \email{douglasrsl@gmail.com}.
}
\keyword{datasets}
