% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_venues.R
\name{search_venues}
\alias{search_venues}
\alias{search_venue}
\title{Search Venues on SetListFM API.}
\usage{
search_venue(search_term, search_by = "cityName", key = return_key())
}
\arguments{
\item{search_term}{Term to be searched}

\item{search_by}{Parameter to be searched - Options are: cityId, cityName, country, name, state stateCode}

\item{key}{API key.}
}
\value{
\code{search_venues} returns a tibble with venues according to the query parameters
}
\description{
Returns venues from SetlistFM API.
}
\examples{

\dontrun{
key <- "your token"
artist <- "the-beatles"
get_request(key = key,
artist = artist)
}

}
\author{
Douglas Zickuhr, \email{douglasrsl@gmail.com}.
}
